+++ Control/Joint/Effects/Reader.hs
@@ -1,7 +1,8 @@
 module Control.Joint.Effects.Reader where
 
-import Control.Joint.Abilities (Interpreted (Primary, run)
-	, Transformer (Schema, embed, build, unite), (:>) (T), Liftable)
+import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
+import Control.Joint.Abilities.Transformer (Transformer (Schema, embed, build, unite), (:>) (T))
+import Control.Joint.Abilities.Liftable (Liftable (lift))
 import Control.Joint.Schemes (TU (TU))
 
 newtype Reader e a = Reader (e -> a)
@@ -36,10 +37,10 @@ instance Applicative u => Applicative (TU ((->) e) u) where
 instance (Applicative u, Monad u) => Monad (TU ((->) e) u) where
 	TU x >>= f = TU $ \e -> x e >>= ($ e) . run . f
 
-get :: Reader e e
-get = Reader $ \e -> e
+type Configured e = Liftable (Reader e)
+
+get :: Configured e t => t e
+get = lift $ Reader $ \e -> e
 
 local :: (e -> i) -> Reader i a -> Reader e a
 local g (Reader f) = Reader (f . g)
-
-type Configured e = Liftable (Reader e)